feat(e2e): implement comprehensive accessibility, performance, and structure tests for the home page
- Added accessibility tests to ensure keyboard navigation and image alt attributes are present.
- Implemented performance tests to verify the home page loads without errors.
- Created structure tests to check for essential semantic elements, proper landmarks, interactive elements, and heading hierarchy.
- Introduced utility functions for base URL and viewport sizes to streamline test setup.

Files changed:
- e2e/home-accessibility.spec.ts
- e2e/home-performance.spec.ts
- e2e/home-responsive.spec.ts
- e2e/home-structure.spec.ts
- e2e/test-utils.ts
