+++ e2e/home-accessibility.spec.ts
@@ -1,0 +1,33 @@
+import { test, expect } from '@playwright/test'
+import { BASE_URL } from './test-utils'
+
+test.describe('Home Page Accessibility', () => {
+  test('is keyboard accessible', async ({ page }) => {
+    await page.goto(BASE_URL)
+
+    // Test that focusable elements exist
+    const focusableElements = page.locator(
+      'a, button, input, select, textarea, [tabindex]:not([tabindex="-1"])'
+    )
+    await expect(focusableElements.first()).toBeVisible()
+
+    // Test tab navigation by focusing first element and checking it receives focus
+    const firstFocusable = focusableElements.first()
+    await firstFocusable.focus()
+    await expect(firstFocusable).toBeFocused()
+  })
+
+  test('images have accessibility attributes', async ({ page }) => {
+    await page.goto(BASE_URL)
+
+    const images = page.locator('img')
+    const imageCount = await images.count()
+
+    if (imageCount > 0) {
+      // Test that all images have alt attributes
+      for (let i = 0; i < imageCount; i++) {
+        await expect(images.nth(i)).toHaveAttribute('alt')
+      }
+    }
+  })
+})
