+++ e2e/test-utils.ts
@@ -1,0 +1,7 @@
+export const BASE_URL = 'http://localhost:4321/'
+
+export const VIEWPORTS = {
+  mobile: { width: 375, height: 667 },
+  tablet: { width: 768, height: 1024 },
+  desktop: { width: 1200, height: 800 },
+}
